1.3.1 - cheditor 5.1.4 업데이트

+++ cheditor5/imageUpload/delete.php
@@ -2,14 +2,19 @@
 
 require_once("config.php");
 // ---------------------------------------------------------------------------
+$filepath = $_POST["filepath"];
+$r = false;
 
-$delete = $_GET['img'];
-$pos = strrpos($delete, '/');
-
-$delete = substr($delete, $pos+1);
-$filepath = sprintf("%s/%s", SAVE_DIR, $delete);
-
-$r = unlink($filepath);
+if (file_exists($filepath)) {
+	$r = unlink($filepath);
+	if ($r) {
+		$thumbPath = dirname($filepath) . DIRECTORY_SEPARATOR . "thumb_" . basename($filepath);
+		if (file_exists($thumbPath)) {
+			unlink($thumbPath);
+		}
+	}
+}
 
 echo $r ? true : false;
+
 ?>
